+++ src/ScottPlot/Drawing/IColormap.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ScottPlot.Drawing
+{
+    public interface IColormap
+    {
+        (byte r, byte g, byte b) GetRGB(byte value);
+    }
+}